Move digital_ocean_v2 tests to be the defauly/only tests for digital ocean cloud tests
