+++ tests/integration/files/conf/cloud.profiles.d/digital_ocean.conf
@@ -1,4 +1,4 @@
 digitalocean-test:
   provider: digitalocean-config
-  image: 14.04 x64
+  image: 14.10 x64
   size: 2GB
