--- tests/integration/files/conf/cloud.profiles.d/digital_ocean_v2.conf
@@ -1,6 +1,0 @@
-digitalocean-test:
-    provider: digitalocean-config
-    image: 6372321
-    size: 512MB
-    location: New York 2
-    private_networking: True
