+++ tests/integration/files/conf/cloud.providers.d/digital_ocean.conf
@@ -1,7 +1,6 @@
 digitalocean-config:
   provider: digital_ocean
-  client_key: ''
-  api_key: ''
+  personal_access_token: ''
   ssh_key_file: ''
   ssh_key_name: ''
   location: New York 1
